fix(typeUtils): export typeUtils and move them to helpers
