+++ packages/patternfly-4/react-core/src/components/ApplicationLauncher/ApplicationLauncher.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps, ReactNode } from 'react';
-import { OneOf } from '../../typeUtils';
+import { OneOf } from '../../helpers/typeUtils';
 import { ApplicationLauncherDirection, ApplicationLauncherPosition } from './applicationLauncherConstants';
 
 export interface ApplicationLauncherProps extends HTMLProps<HTMLDivElement> {
