+++ packages/patternfly-4/react-core/src/components/BackgroundImage/BackgroundImage.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps } from 'react';
-import { Omit } from '../../typeUtils';
+import { Omit } from '../../helpers/typeUtils';
 
 export const BackgroundImageSrc: {
   xs: 'xs';
