+++ packages/patternfly-4/react-core/src/components/DataList/DataListCheck.d.ts
@@ -1,5 +1,5 @@
 import { FormEvent, FunctionComponent, HTMLProps } from 'react';
-import { Omit } from '../../typeUtils';
+import { Omit } from '../../helpers/typeUtils';
 
 export interface DataListCheckProps
   extends Omit<HTMLProps<HTMLInputElement>, 'type' | 'onChange' | 'disabled' | 'aria-labelledby'> {
