+++ packages/patternfly-4/react-core/src/components/DataList/DataListContent.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps } from 'react';
-import { Omit } from '../../typeUtils';
+import { Omit } from '../../helpers/typeUtils';
 
 export interface DataListContentProps extends Omit<HTMLProps<HTMLElement>, 'aria-label'> {
   isHidden: boolean;
