+++ packages/patternfly-4/react-core/src/components/DataList/DataListToggle.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps } from 'react';
-import { Omit } from '../../typeUtils';
+import { Omit } from '../../helpers/typeUtils';
 
 export interface DataListToggleProps extends Omit<HTMLProps<HTMLDivElement>, 'aria-labelledby' | 'aria-label' | 'id'> {
   isExpanded: boolean;
