+++ packages/patternfly-4/react-core/src/components/Nav/NavList.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps, ReactNode } from 'react';
-import { OneOf } from '../../typeUtils';
+import { OneOf } from '../../helpers/typeUtils';
 import { NavVariants } from './NavVariants';
 
 export interface NavListProps extends HTMLProps<HTMLDivElement> {
