+++ packages/patternfly-4/react-core/src/components/Page/PageSection.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps, ReactNode } from 'react';
-import { OneOf } from '../../typeUtils';
+import { OneOf } from '../../helpers/typeUtils';
 
 export const PageSectionVariants: {
   default: 'default';
