+++ packages/patternfly-4/react-core/src/components/Progress/Progress.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps, ReactNode } from 'react';
-import { OneOf, Omit } from '../../typeUtils';
+import { OneOf, Omit } from '../../helpers/typeUtils';
 
 export const ProgressSize: {
   sm: 'sm',
