+++ packages/patternfly-4/react-core/src/components/TextArea/TextArea.d.ts
@@ -1,5 +1,5 @@
 import { FormEvent, FunctionComponent, HTMLProps } from 'react';
-import { Omit } from '../../typeUtils';
+import { Omit } from '../../helpers/typeUtils';
 
 export interface TextAreaProps extends Omit<HTMLProps<HTMLInputElement>, 'onChange'> {
   isRequired?: boolean;
