+++ packages/patternfly-4/react-core/src/components/Tooltip/Tooltip.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps, ReactElement } from 'react';
-import { Omit } from '../../typeUtils';
+import { Omit } from '../../helpers/typeUtils';
 import { BasicPlacement, Props } from 'tippy.js';
 
 export const TooltipPosition: {
