+++ packages/patternfly-4/react-core/src/helpers/index.ts
@@ -2,3 +2,4 @@ export { default as GenerateId } from './GenerateId/GenerateId';
 export * from './util';
 export * from './constants';
 export * from './htmlConstants';
+export * from './typeUtils';
