+++ packages/patternfly-4/react-core/src/layouts/Grid/Grid.d.ts
@@ -1,6 +1,6 @@
 import { HTMLProps, ComponentType, ClassAttributes } from 'react';
 import { GutterSize } from '../../styles/gutters';
-import { OneOf } from '../../typeUtils';
+import { OneOf } from '../../helpers/typeUtils';
 
 export interface GridProps extends HTMLProps<HTMLDivElement> {
   gutter?: OneOf<typeof GutterSize, keyof typeof GutterSize>;
