+++ packages/patternfly-4/react-core/src/layouts/Grid/gridUtils.d.ts
@@ -1,4 +1,4 @@
-import { OneOf } from '../../typeUtils';
+import { OneOf } from '../../helpers/typeUtils';
 import { DeviceSizes } from '../../styles/sizes';
 
 export const gridSpans: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
