+++ packages/patternfly-4/react-core/src/layouts/Level/Level.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps } from 'react';
-import { OneOf } from '../../typeUtils';
+import { OneOf } from '../../helpers/typeUtils';
 import { GutterSize } from '../../styles/gutters';
 
 export interface LevelProps extends HTMLProps<HTMLDivElement> {
