+++ packages/patternfly-4/react-core/src/layouts/Stack/Stack.d.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, HTMLProps, ReactType } from 'react';
-import { OneOf } from '../../typeUtils';
+import { OneOf } from '../../helpers/typeUtils';
 import { GutterSize } from '../../styles/gutters';
 
 export interface StackProps extends HTMLProps<HTMLDivElement> {
